+++ __tests__/helpers/get-api-mock.ts
@@ -2,7 +2,6 @@ import nock from 'nock';
 
 import config from '../../src/lib/config.js';
 import { getUserAgent } from '../../src/lib/readmeAPIFetch.js';
-import readmeAPIv2Oas from '../../src/lib/types/openapiDoc.js';
 
 import { mockVersion } from './oclif.js';
 
@@ -43,8 +42,3 @@ export function getAPIv2MockForGHA(reqHeaders: nock.Options['reqheaders'] = {}) 
     ...reqHeaders,
   });
 }
-
-/**
- * Mocks the fetch call to ReadMe API v2's OpenAPI spec.
- */
-export const oasFetchMock = (status = 200) => getAPIv2Mock().get('/openapi.json').reply(status, readmeAPIv2Oas);